third_party/WebKit/LayoutTests/fast/events/wheel/wheel-in-scrollbar-expected.txt

Test that wheel scrolling over the page scrollbar scrolls the page. We also placed an overflow div below the scrollbar to make sure we target the page and not the div in that case.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.scrollY is 30
PASS overflowDiv.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
